requirements.txt
anyio==4.3.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blessed==1.20.0
bs4==0.0.2
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
editor==1.6.6
getch==1.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==2.10
instaloader==4.10.3
keyboard==0.13.5
mailtrap==2.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
pyarmor==8.4.7
pyarmor.cli.core==5.4.3
pyarmor.cli.core.android==5.4.3
pycryptodome==3.20.0
pyfiglet==1.0.2
Pygments==2.17.2
readchar==4.0.5
requests==2.31.0
rich==13.7.0
runs==1.2.2
scapy==2.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.2
urllib3==1.26.18
wcwidth==0.2.13
xmod==1.8.1
